Extract component tests.
